# Label values with matchers by intersecting postings