@colega 2021-12-16
@pstibrany suggested me to run benchmarks with count>1, so I did with count=10 and here are the results (I've skipped the results of the intermediate implementation this time):

```
name                                                                   old time/op    new time/op    delta
Querier/Head/labelValuesWithMatchers/i_with_n="1"                         140ms ± 0%     107ms ± 2%   -23.67%  (p=0.000 n=9+9)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"                      1.63s ± 6%     0.18s ± 3%   -89.08%  (p=0.000 n=9+9)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"                129ms ± 2%     125ms ± 9%    -2.91%  (p=0.035 n=10+10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"      275ms ± 3%     184ms ± 8%   -33.03%  (p=0.000 n=10+10)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"                      676ms ± 0%     130ms ± 0%   -80.72%  (p=0.000 n=10+9)
Querier/Head/labelValuesWithMatchers/n_with_i="1"                        5.35µs ± 1%    4.43µs ± 5%   -17.28%  (p=0.000 n=9+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1"                        109ms ± 1%     111ms ± 0%    +2.56%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"                     1.37s ± 1%     0.15s ± 0%   -89.30%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"              88.7ms ± 1%   117.9ms ± 1%   +32.83%  (p=0.000 n=9+9)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"     177ms ± 1%     163ms ± 1%    -7.92%  (p=0.000 n=9+9)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"                     1.14s ± 1%     0.14s ± 2%   -87.85%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/n_with_i="1"                        973µs ± 6%     928µs ± 3%    -4.68%  (p=0.001 n=9+9)

name                                                                   old alloc/op   new alloc/op   delta
Querier/Head/labelValuesWithMatchers/i_with_n="1"                        29.5MB ± 0%    19.1MB ± 0%   -35.36%  (p=0.000 n=9+10)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"                     46.8MB ± 0%    36.3MB ± 0%   -22.30%  (p=0.000 n=10+10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"               29.5MB ± 0%    19.1MB ± 0%   -35.37%  (p=0.000 n=9+10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"     46.8MB ± 0%    20.7MB ± 0%   -55.79%  (p=0.000 n=10+10)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"                     10.3kB ± 2%     6.4kB ± 0%   -37.99%  (p=0.000 n=10+10)
Querier/Head/labelValuesWithMatchers/n_with_i="1"                        5.54kB ± 0%    4.30kB ± 0%   -22.51%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1"                       39.1MB ± 0%    20.7MB ± 0%   -47.14%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"                     287MB ± 0%      38MB ± 0%   -86.77%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"              34.3MB ± 0%    20.7MB ± 0%   -39.75%  (p=0.000 n=9+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"    51.6MB ± 0%    22.3MB ± 0%   -56.80%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"                     144MB ± 0%       0MB ± 0%   -99.99%  (p=0.000 n=7+10)
Querier/Block/labelValuesWithMatchers/n_with_i="1"                       6.43kB ± 0%    5.86kB ± 0%    -8.83%  (p=0.000 n=10+10)

name                                                                   old allocs/op  new allocs/op  delta
Querier/Head/labelValuesWithMatchers/i_with_n="1"                          104k ± 0%      300k ± 0%  +188.57%  (p=0.000 n=9+10)
Querier/Head/labelValuesWithMatchers/i_with_n="^.+$"                       204k ± 0%      400k ± 0%   +96.11%  (p=0.000 n=10+10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",j!="foo"                 104k ± 0%      300k ± 0%  +188.51%  (p=0.000 n=9+10)
Querier/Head/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"       204k ± 0%      300k ± 0%   +47.06%  (p=0.000 n=10+10)
Querier/Head/labelValuesWithMatchers/n_with_j!="foo"                       66.5 ± 1%     139.0 ± 0%  +109.02%  (p=0.000 n=10+10)
Querier/Head/labelValuesWithMatchers/n_with_i="1"                          61.0 ± 0%      87.0 ± 0%   +42.62%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1"                         304k ± 0%      400k ± 0%   +31.60%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="^.+$"                     5.20M ± 0%     0.50M ± 0%   -90.39%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1",j!="foo"                204k ± 0%      400k ± 0%   +96.11%  (p=0.000 n=9+10)
Querier/Block/labelValuesWithMatchers/i_with_n="1",i=~"^.*$",j!="foo"      304k ± 0%      400k ± 0%   +31.58%  (p=0.000 n=10+10)
Querier/Block/labelValuesWithMatchers/n_with_j!="foo"                     3.00M ± 0%     0.00M ± 0%   -99.99%  (p=0.000 n=7+10)
Querier/Block/labelValuesWithMatchers/n_with_i="1"                         61.0 ± 0%     129.0 ± 0%  +111.48%  (p=0.000 n=10+10)
```